% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casnet_SOURCE_jamovi.R
\name{as.numeric_discrete}
\alias{as.numeric_discrete}
\title{Discrete (factor or character) to numeric vector}
\usage{
as.numeric_discrete(x, keepNA = FALSE, sortUnique = FALSE)
}
\arguments{
\item{x}{A factor with levels that are numeric, or, a character vector representing numbers.}

\item{keepNA}{Keep NA values (`TRUE`), or remove them (default = `FALSE`)}

\item{sortUnique}{Should the unique character/factor level values be sorted? (default = `FALSE`)}
}
\value{
A numeric vector with factor levels / numeric character values as names.
}
\description{
Converts a factor with numeric levels, or, a character vector with numeric values to a numeric vector using [as.numeric_factor], or, [as.numeric_character] respectively. If an unnamed numeric vector is passed, it will be returned as a named numeric vector.
}
\examples{

# Continuous
i <- runif(10,0,9)
as.numeric_discrete(i)

# Integer
as.numeric_discrete(round(i))

# Factor with NAs
f <- factor(c(round(runif(9,0,9)),NA))
as.numeric_discrete(f)
as.numeric_discrete(f, keepNA = FALSE)

# Character vector
c <- c("Thank","you", "for", "the flowers")
as.numeric_discrete(c)
as.numeric_discrete(c, sortUnique = TRUE)

c <- c("Thank","you", "for", "the", "flowers")
as.numeric_discrete(c)
as.numeric_discrete(c, sortUnique = TRUE)


}
