% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casnet-package.r
\docType{package}
\name{casnetjmv-package}
\alias{casnetjmv}
\alias{casnetjmv-package}
\title{casnetjmv: CASNET}
\description{
JAMOVI Module. A collection of analytic tools for studying signals recorded from complex adaptive systems or networks: (Cross) Recurrence Quantification Analyses (Continuous and Categorical (C)RQA, Chromatic RQA, Anisotropic RQA). Fluctuation Analyses (DFA, PSD slope, SDA, Multifractal DFA, Wavelet Singularity Spectrum). Network based time series analyses and visualisation (Recurrence Networks, spiral graphs). Multivariate (network based) time series analyses (Dynamic Complexity, Multiplex Recurrence Networks).
}
\author{
\strong{Maintainer}: Fred Hasselman \email{fred.hasselman@ru.nl} (\href{https://orcid.org/0000-0003-1384-8361}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Merlijn Olthof \email{m.olthof@pwo.ru.nl} [contributor]
}

}
\keyword{internal}
