% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casnet_SOURCE_jamovi.R
\name{di2bi}
\alias{di2bi}
\title{Distance to binary matrix}
\usage{
di2bi(distmat, emRad, theiler = 0, convMat = FALSE)
}
\arguments{
\item{distmat}{Distance matrix}

\item{emRad}{The radius or threshold value}

\item{theiler}{= Use a theiler window around the line of identity / synchronisation to remove high auto-correlation at short time-lags (default = `0`)}

\item{convMat}{Should the matrix be converted from a `distmat` obkect of class [Matrix::Matrix()] to [base::matrix()] (or vice versa)}
}
\value{
A (sparse) matrix with only 0s and 1s
}
\description{
Distance matrix to binary matrix based on threshold value
}
\seealso{
Other Distance matrix operations (recurrence plot): \code{\link{di2we}},
  \code{\link{rp_plot_jmv}}

Other Distance matrix operations (recurrence network): \code{\link{di2we}},
  \code{\link{rn_plot}}, \code{\link{rn_recSpec}},
  \code{\link{rn_scaleoGram}}, \code{\link{rn}}
}
\concept{Distance matrix operations (recurrence network)}
\concept{Distance matrix operations (recurrence plot)}
