% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casnet_SOURCE_jamovi.R
\name{di2we}
\alias{di2we}
\title{Distance 2 weighted matrix}
\usage{
di2we(distmat, emRad, convMat = FALSE)
}
\arguments{
\item{distmat}{Distance matrix}

\item{emRad}{The radius or threshold value}

\item{convMat}{convMat Should the matrix be converted from a `distmat` obkect of class [Matrix::Matrix()] to [base::matrix()] (or vice versa)}
}
\value{
A matrix with 0s and values < threshold distance value
}
\description{
Distance matrix to weighted matrix based on threshold value
}
\seealso{
Other Distance matrix operations (recurrence plot): \code{\link{di2bi}},
  \code{\link{rp_plot_jmv}}

Other Distance matrix operations (recurrence network): \code{\link{di2bi}},
  \code{\link{rn_plot}}, \code{\link{rn_recSpec}},
  \code{\link{rn_scaleoGram}}, \code{\link{rn}}
}
\concept{Distance matrix operations (recurrence network)}
\concept{Distance matrix operations (recurrence plot)}
