% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casnet_SOURCE_jamovi.R
\name{fd_allan}
\alias{fd_allan}
\title{Allan Variance Analysis}
\usage{
fd_allan(y, fs = stats::tsp(stats::hasTsp(y))[3], doPlot = FALSE,
  useSD = FALSE)
}
\arguments{
\item{y}{A numeric vector or time series object}

\item{fs}{Sample frequency in Hz}

\item{doPlot}{Produce a plot of the Allan variance?}

\item{useSD}{Use the standarddeviation instead of variance?}
}
\value{
A dataframe with for each scale \code{tau}, the Allan variance (or Allan factor for poisson processes), Alan standard deviation and error due to bin size
}
\description{
Allan Variance Analysis
}
