% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casnet_SOURCE_jamovi.R
\name{fltrIT}
\alias{fltrIT}
\title{Wrapper for filtfilt}
\usage{
fltrIT(TS, f)
}
\arguments{
\item{TS}{A time series}

\item{f}{A filter}
}
\value{
A filtered signal
}
\description{
Wrapper for filtfilt
}
\keyword{internal}
