% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casnet_SOURCE_jamovi.R, R/invictor_jamovi.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Whole number?}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)

is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{A number}

\item{tol}{The tolerance (\code{default = Machine$double.eps^0.5})}

\item{x}{Number to check}

\item{tol}{Tolerance}
}
\value{
TRUE or FALSE

TRUE if \code{x} is a signed integer.
}
\description{
Whole number?

Wholenumber check
}
\note{
This code was found in the examples of \link[base]{is.integer}.
}
\keyword{internal}
