% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casnet_SOURCE_jamovi.R
\name{layout_as_spiral}
\alias{layout_as_spiral}
\title{Layout a graph on a spiral}
\usage{
layout_as_spiral(g, type = c("Archimedean", "Bernoulli", "Fermat",
  "Euler"), arcs = 6, a = 1, b = NULL, rev = FALSE)
}
\arguments{
\item{g}{An igraph object. If (`rev = FALSE`) the vertex with the lowest index will be placed in the centre of the spiral, the highest index will be most outer vertex,}

\item{type}{Spiral type, one of `"Archimedean"`,`"Bernoulli"`,`"Fermat"`, or, `"Euler"` (default = `"Archimedean"`)}

\item{arcs}{The number of arcs (half circles/ovals) that make up the spiral (default = `10`)}

\item{a}{Parameter controlling the distance between spiral arms, however, the effect will vary for different spiral types (default = `0.5`)}

\item{b}{Parameter controlling where the spiral originates. A value of 1 will generally place the origin in the center. The default `NULL` will choose a value based on the different spiral types (default = `NULL`)}

\item{rev}{If `TRUE` the vertex with the highest index will be placed in the centre of the spiral (default = `FALSE`)}
}
\value{
An igraph layout
}
\description{
Layout a graph on a spiral
}
\examples{

library(igraph)

g  <- igraph::sample_gnp(100, 1/100)

# Equiangular spiral: Any line from the origin cuts at the same angle.
plot(g, layout = layout_as_spiral(g, type = "Bernoulli", arcs = 5))

# The arms of Fermat's spiral diverge quadratically.
plot(g, layout = layout_as_spiral(g, type = "Fermat", arcs = 5))

# Equidistance of intersection points along a line through the origin.
plot(g, layout = layout_as_spiral(g, type = "Archimedean", arcs = 5))

}
