% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casnet_SOURCE_jamovi.R
\name{make_spiral_graph}
\alias{make_spiral_graph}
\title{Make Spiral Graph}
\usage{
make_spiral_graph(g, type = "Archimedean", arcs = 6, a = 1,
  b = NULL, rev = FALSE, curvature = -0.6, angle = 90,
  markTimeBy = NULL, alphaV = 1, alphaE = 0.6, title = "",
  subtitle = "", showEpochLegend = TRUE, markEpochsBy = NULL,
  epochColours = NULL, epochLabel = "Epoch", showSizeLegend = FALSE,
  sizeLabel = "Size", scaleVertexSize = c(1, 6),
  vertexBorderColour = "black", scaleEdgeSize = 1/5,
  edgeColourLabel = "Weight", showEdgeColourLegend = FALSE,
  edgeColourByEpoch = TRUE, defaultEdgeColour = "grey70",
  doPlot = TRUE)
}
\arguments{
\item{g}{An igraph object. If (`rev = FALSE`) the vertex with the lowest index will be placed in the centre of the spiral, the highest index will be most outer vertex,}

\item{type}{Spiral type, one of `"Archimedean"`,`"Bernoulli"`,`"Fermat"`, or, `"Euler"` (default = `"Archimedean"`)}

\item{arcs}{The number of arcs (half circles/ovals) that make up the spiral (default = `10`)}

\item{a}{Parameter controlling the distance between spiral arms, however, the effect will vary for different spiral types (default = `0.5`)}

\item{b}{Parameter controlling where the spiral originates. A value of 1 will generally place the origin in the center. The default `NULL` will choose a value based on the different spiral types (default = `NULL`)}

\item{rev}{If `TRUE` the vertex with the highest index will be placed in the centre of the spiral (default = `FALSE`)}

\item{curvature}{The `curvature` parameter for edges see [geom_curve()] (default = `-0.7`)}

\item{angle}{The `angle` parameter for edges see [geom_curve()] (default = `90`)}

\item{markTimeBy}{Include a vector that indicates time. The time will be displayed on the plot. Pass `TRUE` to generate auto labels (experimental)}

\item{title}{A title for the plot}

\item{subtitle}{A subtitle for the plot}

\item{showEpochLegend}{Should a legend be shown for the epoch colours? (default = `TRUE`)}

\item{markEpochsBy}{A vector of length `vcount(g)` indicating epochs or groups (default = `NULL`)}

\item{epochColours}{A vector of length `vcount(g)` with colour codes (default = `NULL`)}

\item{epochLabel}{A title for the epoch legend (default = `"Epoch"`)}

\item{showSizeLegend}{Should a legend be shown for the size of the nodes? (default = `FALSE`)}

\item{sizeLabel}{Use to indicate if `V(g)$size` represents some measure, e.g. [igraph::degree()], or, [igraph::hubscore()] (default = `"Size"`)}

\item{scaleVertexSize}{Scale the size of the vertices by setting a range for [ggplot2::scale_size()]. This will not affect the numbers on the size legend (default = `c(1,6)`)}

\item{vertexBorderColour}{Draw a border around the vertices. Pass `NULL` to use the same colour as the fill colour (default = `"black"`)}

\item{scaleEdgeSize}{Scale the size of the edges by a constant: `E(g)$width * scaleEdgeSize` (default = `1/5`)}

\item{edgeColourLabel}{Use to indicate if `E(g)$color` represents color coding based on some property. (default = `"Weight"`)}

\item{showEdgeColourLegend}{Should a legend be shown for the colour of the edges? (default = `FALSE`)}

\item{edgeColourByEpoch}{Should edges that connect to the same epoch be assigned the epoch colour? This will ignore edge colour info in `E(g)$color`. (default = `TRUE`)}

\item{defaultEdgeColour}{Colour of edges that do not connect to the same epoch (default = `"grey70"`)}

\item{doPlot}{Produce a plot? (default = `TRUE`)}
}
\value{
A ggplot object.
}
\description{
Turn an [igraph] object into a spiral graph returning a [ggplot2] object.
}
\note{
To keep the igraph object, use the layout function [layout_as_spiral(g)] when plotting the graph.
}
\examples{

library(igraph)

g  <- igraph::sample_gnp(200, 1/20)
V(g)$size <- degree(g)
make_spiral_graph(g, markTimeBy = TRUE, showSizeLegend = TRUE, sizeLabel = "Node degree")

}
