% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFUN.R
\name{pacf_fisherZ}
\alias{pacf_fisherZ}
\title{#' FisherZ for PACF
#'
#' @param r r
#' @param n n
#' @param lag lag
#' @param siglevel siglevel
#' @param sides sides
#' @param type type
#'
#' @return dataframe
#' @export
#'
#' @keywords internal
#'
pacf_fisherZ <-function(r, n, lag, siglevel=.05,sides=2,type=""){
z <- atanh(r)/sqrt(1/(n-3))
conf.low   <- tanh(atanh(r) - (stats::qnorm(1-(siglevel/sides)) * sqrt((1/(n-3)))))
conf.high  <- tanh(atanh(r) + (stats::qnorm(1-(siglevel/sides)) * sqrt((1/(n-3)))))
p<-2*(1-stats::pnorm(abs(z)))
if(p<siglevel){
sig <-"yes"
} else {
sig <-"no"
}
return(data.frame(r=r,ciL=conf.low,ciU=conf.high,fZ=z,p=p,alpha=siglevel,sig=sig,lag=lag,n=n,type=type, stringsAsFactors = FALSE))
}
FisherZ for PACF}
\usage{
pacf_fisherZ(
  r,
  n,
  lag,
  siglevel = 0.05,
  sides = 2,
  alt = "both",
  type = "",
  label = ""
)
}
\arguments{
\item{r}{r}

\item{n}{n}

\item{lag}{lag}

\item{siglevel}{siglevel}

\item{sides}{sides}

\item{alt}{alt}

\item{type}{type}
}
\value{
dataframe
}
\description{
#' FisherZ for PACF
#'
#' @param r r
#' @param n n
#' @param lag lag
#' @param siglevel siglevel
#' @param sides sides
#' @param type type
#'
#' @return dataframe
#' @export
#'
#' @keywords internal
#'
pacf_fisherZ <-function(r, n, lag, siglevel=.05,sides=2,type=""){
z <- atanh(r)/sqrt(1/(n-3))
conf.low   <- tanh(atanh(r) - (stats::qnorm(1-(siglevel/sides)) * sqrt((1/(n-3)))))
conf.high  <- tanh(atanh(r) + (stats::qnorm(1-(siglevel/sides)) * sqrt((1/(n-3)))))
p<-2*(1-stats::pnorm(abs(z)))
if(p<siglevel){
sig <-"yes"
} else {
sig <-"no"
}
return(data.frame(r=r,ciL=conf.low,ciU=conf.high,fZ=z,p=p,alpha=siglevel,sig=sig,lag=lag,n=n,type=type, stringsAsFactors = FALSE))
}
FisherZ for PACF
}
