% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFUN.R
\name{plotFA_loglog}
\alias{plotFA_loglog}
\title{plotFA_loglog}
\usage{
plotFA_loglog(
  fd.OUT,
  title = "log-log regresion",
  subtitle = "",
  xlabel = "Bin size",
  ylabel = "Fluctuation",
  logBase = c("e", "2", "10")[1]
)
}
\arguments{
\item{fd.OUT}{y}

\item{title}{y}

\item{subtitle}{y}

\item{xlabel}{y}

\item{ylabel}{y}

\item{logBase}{y}
}
\description{
plotFA_loglog
}
