% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casnet_SOURCE_jamovi.R
\name{plotNET_groupColour}
\alias{plotNET_groupColour}
\title{Vertex and Edge Group Colours}
\usage{
plotNET_groupColour(g, groups, colourV = TRUE, alphaV = 1,
  colourE = FALSE, alphaE = 0.8, groupColours = NULL,
  defaultEdgeColour = "grey70", doPlot = TRUE)
}
\arguments{
\item{g}{An igraph object}

\item{groups}{A named numeric vector with `length(V(g))` integers representing each group, or, a named character vector describing each group. If `names(groups)==NULL` then the names of the vector will be set as `names(groups) == V(g)$name`. If `V(g)$name==NULL`, the names of the vector will be set by the Vertex index}

\item{colourV}{Colour Vertices based on `groups` (default = `TRUE`)}

\item{alphaV}{Set transparency for Vertices (default = `1`)}

\item{colourE}{Colour Edges based on `groups`. Edges connecting to vertices of the same group will be coloured as the group (default = `FALSE`)}

\item{alphaE}{Set transparency for Edges. A single numeric, or a vector of length `ecount(g)` (default = `0.8`)}

\item{groupColours}{A list of length `groups` with valid colour codes}

\item{defaultEdgeColour}{Default edge colour}

\item{doPlot}{Plot the igraph object}
}
\value{
An igraph object with vertices and/or edges coloured by groups listed in `groups`
}
\description{
Identify Vertex and/or Edge groups by colour.
}
\seealso{
Other tools for plotting networks: \code{\link{plotNET_BA}},
  \code{\link{plotNET_SW}},
  \code{\link{plotNET_groupWeight}},
  \code{\link{plotNET_prep}}
}
\concept{tools for plotting networks}
