% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casnet_SOURCE_jamovi.R
\name{pacf_fisherZ}
\alias{pacf_fisherZ}
\title{FisherZ for PACF}
\usage{
pacf_fisherZ(r, n, lag, siglevel = 0.05, sides = 2, alt = "both",
  type = "", label = "")
}
\arguments{
\item{r}{r}

\item{n}{n}

\item{lag}{lag}

\item{siglevel}{siglevel}

\item{sides}{sides}

\item{alt}{alt}

\item{type}{type}
}
\value{
dataframe
}
\description{
FisherZ for PACF
}
