% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casnet_SOURCE_jamovi.R
\name{ssg_winnowing}
\alias{ssg_winnowing}
\title{ssg_winnowing}
\usage{
ssg_winnowing(durations, screeCut, timeLimit = 120)
}
\arguments{
\item{durations}{durations frame}

\item{screeCut}{cutoff}

\item{timeLimit}{Stop the procedure after this time limit isn seconds (default = `120`)}
}
\value{
attractor frame
}
\description{
ssg_winnowing
}
